main update and binaries remove
