+++ app.ignite.yml
@@ -1,6 +1,6 @@
 version: 1
 apps:
-  ignite-notify:
+  notify:
     description: ignite-notify is a modular notification service designed for the Ignite blockchain ecosystem. It could monitors on-chain events, validator activities, governance proposals, and transaction statuses,
   then delivers real-time notifications to users through multiple channels such as email, Telegram,
   and webhooks. The application supports configurable notification rules, and
